feat(execute_withdrawal): Renamed field_element_utils to field_element
